Final rename for release 1.0.  Now, the tame keywords are just:
	tamed, twait, and tvars

Tame classes are just:

	event_t,rendezvous_t

Make new events with 'mkevent' and not @().


To convert over okcupid, we're going to have run a script to convert
everything.  But let's put this off for a while...

The advantage is that it is much cleaner looking, and a lot more of the 
code is in the library (as opposed to in the translator).  the disadvantage
is that it's going to be a pain to move everything over, and might not
be worth it. 

it's also possible to temporarily reinsert keywords for backwards
compatibility.



git-svn-id: svn://okws.org/sfslite2/devel/concur-0-1@2272 7d287422-6eea-da11-b677-00123f2a783a