Bindu Pucha and Michael Kaminsky submitted these patches:
I have two more patches for you (attached).  One is against axprt_pipe,
that adds an one-line accessor function to get the number of bytes in
the output buffer.
 
Two is a little more involved.  It's a patch against aiod to add
directory functions (opendir, readdir, closedir).  It seems like an
obvious feature to have in aiod.  The patch is Bindu's (except a bit a
cleanup I did)..



git-svn-id: svn://okws.org/sfslite2/devel/1.1@2881 7d287422-6eea-da11-b677-00123f2a783a